Update `generic-service` to include grafana dashboards out of the box
This change allows all users of this chart to get a "Golden Signals" dashboard for their service - showing HTTP req rate, latency and errors, as well as CPU/memory use.

If the service has an ingress the dashboard will be based on metrics from the NGINX ingress controllers.

If the service does not have an ingress the following will happen:

- A traefik proxy will be injected into the deployment in front of the main container. This is there to provide request metrics in absence of an ingress.
- A network policy will be added to allow metrics to be collected from traefik.
- A service monitor will be added to instruct prometheus to scrape traefik metrics.