Add a CronJobStatusFailed alert
This adds two recording rules to figure out the last ran time of a
cronjob, and it's success/failure state, then uses those rules in the
alert.

This is a generic (whole cluster) setup at the moment and might be a
touch fragile given the size of our cluster and the fact that we have
one prometheus...  I'll scope this to just the namespace in a sec.

ref: https://www.giffgaff.io/tech/monitoring-kubernetes-jobs