Support active record in postgres restore script
We maintain a mixed set of HMPPS applications, some of them are legacy and still using Rails. We are already using the db restore job with kotlin/flyway apps but would be great to also be able to use it in our Rails apps, where we use this helm chart successfully.

This PR allows supplying an optional `MIGRATIONS_VENDOR` env variable to `postgresDatabaseRestore`. If not provided, current behaviour (flyway) remains intact, thus the minor patch version.

Will be very easy to expand this to other migration types, if necessary in the future.