PI-2965 Add support for timeZone in scheduled downtime job (#183)
Non-production RDS databases on Cloud Platform are stopped between 10pm-6am UTC, so for most services the time zone should be left as "Etc/UTC". However, for other services that have dependencies with a downtime schedule in a different timezone (e.g. "Europe/London"), then this allows them to ensure the schedules match.

Specifically, this allows probation integration services that connect to the Delius database to ensure the pods are offline while the database is offline (9PM-6AM Europe/London time).