Update data extractor to latest
We had a critical alert from dependabot in the data extractor repository which required us to bump a few dependencies. We've updated the extractor to a new release which we'll now need to deploy in the instances where the data extractor is used. We will remove the old releases shortly.